fix: selected language was overridden by default one (en) for all translation in app
Put i18next init in I18nService constructor instead of `init` function to avoid multiple calls to it as it was the case now.
As App.tsx is a FunctionComponent, it was called each time it should be renderer to reflect a data change, and so i18nService was re-init each time also.

Fix https://github.com/cycle-game/Cycle/issues/58